Move UnsupportedInfluenceException to influence.types, raise exception on unknown mode
