+++ src/pydvl/influence/base_influence_function_model.py
@@ -10,14 +10,6 @@ from ..utils.progress import log_duration
 from .types import BatchType, BlockMapperType, DataLoaderType, InfluenceMode, TensorType
 
 
-class UnsupportedInfluenceModeException(ValueError):
-    def __init__(self, mode: str):
-        super().__init__(
-            f"Provided {mode=} is not supported. Choose one of InfluenceMode.Up "
-            f"and InfluenceMode.Perturbation"
-        )
-
-
 class NotFittedException(ValueError):
     def __init__(self, object_type: Type):
         super().__init__(
