+++ src/pydvl/influence/types.py
@@ -418,9 +418,14 @@ class OperatorGradientComposition(
             The result of the influence computation as dictated by the mode.
         """
         bilinear_form = self.op.as_bilinear_form()
-        if mode is InfluenceMode.Up:
+        if mode == InfluenceMode.Up:
             return bilinear_form.grads_inner_prod(left_batch, right_batch, self.gp)
-        return bilinear_form.mixed_grads_inner_prod(left_batch, right_batch, self.gp)
+        elif mode == InfluenceMode.Perturbation:
+            return bilinear_form.mixed_grads_inner_prod(
+                left_batch, right_batch, self.gp
+            )
+        else:
+            raise UnsupportedInfluenceModeException(mode)
 
     def transformed_grads(self, batch: BatchType):
         r"""
@@ -646,3 +651,11 @@ class BlockMapper(Generic[TensorType, BatchType, OperatorGradientCompositionType
 
 
 BlockMapperType = TypeVar("BlockMapperType", bound=BlockMapper)
+
+
+class UnsupportedInfluenceModeException(ValueError):
+    def __init__(self, mode: str):
+        super().__init__(
+            f"Provided {mode=} is not supported. Choose one of InfluenceMode.Up "
+            f"and InfluenceMode.Perturbation"
+        )
